refactor: clean all clippy warnings — Phases A through I (SOLID/DRY/KISS)
cargo clippy --lib --tests now produces ZERO warnings (was 47 + 2 errors).

Phase A (errors blocking clippy):
- storage/config_sqlite.rs:508 — replace magic 3.14 (approx PI) with 1.5

Phase B (structural — items_after_test_module):
- orchestrator/recording.rs — moved stage_mapping_tests module to end of file
- storage/mod.rs — moved test_utils module to end of file

Phase C (field_reassign_with_default — 16 sites across 6 files):
- commands/config.rs, storage/config_ini.rs (2), storage/config_sqlite.rs (8),
  storage/factory.rs, orchestrator/transcription/tests.rs (2)
- DRY win in config/validation.rs: 9 tests collapsed via assert_single_error()
  helper taking FnOnce mutator + expected field name

Phase D (clone_on_copy — 3 sites):
- orchestrator/state.rs — already fixed in earlier commit
- permissions/checker.rs (2) — use Clone::clone(&p) for explicit trait test

Phase E (useless_vec — 7 sites):
- commands/overlay.rs, learning/normalizer.rs, permissions/macos.rs (2),
  storage/debug_storage.rs (3) — vec![] -> [] arrays where Vec not needed

Phase F (float excessive_precision — 4 sites in audio/level/tests.rs):
- DRY: extracted HASH_SEED_X/HASH_SEED_Y constants (f32-truncated precision)
- New hash_sample(i) helper returns deterministic [-1, 1] PRNG sample
- 3 test sites use the helper directly, semantics preserved

Phase G (single-occurrence misc — 5 sites):
- audio/vad/tests.rs — result.len() == 0 -> result.is_empty()
- audio/level/tests.rs — val >= 0.0 && val <= 1.0 -> (0.0..=1.0).contains(&val)
- error.rs — io::Error::new(Other, msg) -> io::Error::other(msg) (Rust 1.74+)
- overlay_bin/tests.rs — manual range -> (0.5..=1.0).contains(&f)
- overlay_bin/render.rs — #[allow(dead_code)] for Lua-bridge state_name

Phase H (type_complexity — deferred):
- Phase H type aliases NOT applied yet (was in original plan); current build
  no longer triggers warnings, so deferred until they reappear.

Phase I (overlay_bin macros):
- platform/macos/draw.rs — added sel + sel_impl imports (resolves msg_send!)

Results: 741 Rust + 891 frontend + 95 E2E green (1 native-overlay-themes
flaky test pre-existing on macOS, fails identically with or without these
changes; not a regression).

+++ src-tauri/src/orchestrator/tests.rs
@@ -722,7 +722,7 @@ async fn test_queue_pipeline_multiple_items() {
         }
 
         // Stop recording, push to queue
-        let audio = vec![i as u8; 512];
+        let audio = vec![i; 512];
         let queue_len = queue.push(audio).await;
         assert_eq!(queue_len as u8, i + 1);
